Move interactive tutorial components to correct architectural tier (Phase 4b)
Relocate 13 files from docs-retrieval/components/interactive/ to
components/interactive-tutorial/, placing presentation components in
Tier 4 where they belong. This eliminates 10 lateral violations and
swaps 1 vertical violation for a new (deferred) upward dependency from
content-renderer.tsx -> components.

Net result: vertical=3 (unchanged), lateral=21→11, barrel=0.
