+++ src/components/interactive-tutorial/get-content-key.ts
