+++ src/components/interactive-tutorial/input-block.tsx
@@ -10,8 +10,8 @@ import { css } from '@emotion/css';
 import { Button, Input, Checkbox, Field, useStyles2, Alert, Icon, Combobox, type ComboboxOption } from '@grafana/ui';
 import { GrafanaTheme2 } from '@grafana/data';
 import { getDataSourceSrv } from '@grafana/runtime';
-import { useGuideResponsesOptional } from '../../GuideResponseContext';
-import { reportAppInteraction, UserInteraction } from '../../../lib/analytics';
+import { useGuideResponsesOptional } from '../../docs-retrieval';
+import { reportAppInteraction, UserInteraction } from '../../lib/analytics';
 
 /** Props for the InputBlock component */
 export interface InputBlockProps {
