+++ src/components/interactive-tutorial/interactive-conditional.tsx
@@ -9,10 +9,10 @@
  */
 
 import React, { useState, useEffect, useCallback, useRef, useMemo } from 'react';
-import { useInteractiveElements } from '../../../interactive-engine';
-import type { ParsedElement } from '../../html-parser';
-import { testIds } from '../../../constants/testIds';
-import type { ConditionalDisplayMode, ConditionalSectionConfig } from '../../../types/json-guide.types';
+import { useInteractiveElements } from '../../interactive-engine';
+import type { ParsedElement } from '../../docs-retrieval';
+import { testIds } from '../../constants/testIds';
+import type { ConditionalDisplayMode, ConditionalSectionConfig } from '../../types/json-guide.types';
 import { InteractiveSection } from './interactive-section';
 
 export interface InteractiveConditionalProps {
