+++ src/components/interactive-tutorial/interactive-quiz.tsx
@@ -3,9 +3,9 @@ import { css, cx, keyframes } from '@emotion/css';
 import { Button, Icon, useStyles2 } from '@grafana/ui';
 import { GrafanaTheme2 } from '@grafana/data';
 
-import { useStepChecker } from '../../../requirements-manager';
-import { reportAppInteraction, UserInteraction, buildInteractiveStepProperties } from '../../../lib/analytics';
-import { testIds } from '../../../constants/testIds';
+import { useStepChecker } from '../../requirements-manager';
+import { reportAppInteraction, UserInteraction, buildInteractiveStepProperties } from '../../lib/analytics';
+import { testIds } from '../../constants/testIds';
 import { useStandalonePersistence } from './use-standalone-persistence';
 
 // ============ Types ============
