--- src/docs-retrieval/components/interactive-components.tsx
@@ -1,5 +1,0 @@
-// Re-export interactive components
-export * from './interactive';
-
-// Re-export documentation components
-export * from './docs';
