+++ src/validation/architecture.test.ts
@@ -100,7 +100,8 @@ function collectViolations(getViolationKey: (ctx: ResolvedImportContext) => stri
  */
 const ALLOWED_VERTICAL_VIOLATIONS = new Set([
   'docs-retrieval/content-renderer.tsx -> integrations',
-  'docs-retrieval/components/interactive/interactive-step.tsx -> integrations',
+  // Phase 4b: content-renderer imports moved interactive components (Tier 2 -> Tier 4)
+  'docs-retrieval/content-renderer.tsx -> components',
   // Existing cross-domain persistence dependency (legacy coupling)
   'lib/user-storage.ts -> learning-paths',
 ]);
@@ -120,20 +121,11 @@ const ALLOWED_LATERAL_VIOLATIONS = new Set([
   'requirements-manager/step-checker.hook.ts -> interactive-engine',
   // Cluster B: context-engine -> docs-retrieval
   'context-engine/context.service.ts -> docs-retrieval',
-  // docs-retrieval orchestrates interactive/requirements (many touch points)
+  // docs-retrieval cross-engine imports
   'docs-retrieval/content-renderer.tsx -> requirements-manager',
   'docs-retrieval/json-parser.ts -> validation',
   'docs-retrieval/content-fetcher.ts -> validation',
-  'docs-retrieval/components/interactive/interactive-section.tsx -> interactive-engine',
-  'docs-retrieval/components/interactive/interactive-section.tsx -> requirements-manager',
-  'docs-retrieval/components/interactive/interactive-quiz.tsx -> requirements-manager',
-  'docs-retrieval/components/interactive/interactive-multi-step.tsx -> requirements-manager',
-  'docs-retrieval/components/interactive/interactive-multi-step.tsx -> interactive-engine',
-  'docs-retrieval/components/interactive/interactive-guided.tsx -> requirements-manager',
-  'docs-retrieval/components/interactive/interactive-guided.tsx -> interactive-engine',
-  'docs-retrieval/components/interactive/interactive-conditional.tsx -> interactive-engine',
-  'docs-retrieval/components/interactive/interactive-step.tsx -> requirements-manager',
-  'docs-retrieval/components/interactive/interactive-step.tsx -> interactive-engine',
+  // Phase 4b: 10 interactive component lateral entries removed (files moved to components/)
   // Additional pre-existing cross-engine imports uncovered by AST parsing
   'docs-retrieval/learning-journey-helpers.ts -> learning-paths',
   'requirements-manager/requirements-checker.hook.ts -> context-engine',
